fix: change BigInteger to SmallIteger in social_network_id, tag_id, refresh token status_id in all tables

+++ alembic/versions/2025-09-27_43a6f66a8aa3_replace_cheatsheet_id_on_the_first_place.py
@@ -59,7 +59,7 @@ def upgrade() -> None:
     op.create_table(
         "cheatsheet_to_tag_temp",
         sa.Column("cheatsheet_id", postgresql.UUID(as_uuid=True), nullable=False),
-        sa.Column("tag_id", sa.BigInteger(), nullable=False),
+        sa.Column("tag_id", sa.SmallInteger(), nullable=False),
     )
 
     op.execute(
@@ -193,7 +193,7 @@ def downgrade() -> None:
 
     op.create_table(
         "cheatsheet_to_tag_old",
-        sa.Column("tag_id", sa.BigInteger(), nullable=False),
+        sa.Column("tag_id", sa.SmallInteger(), nullable=False),
         sa.Column("cheatsheet_id", postgresql.UUID(as_uuid=True), nullable=False),
     )
