+++ alembic/versions/2025-11-20_bd475c306381_create_refresh_token_status_model.py
@@ -26,7 +26,7 @@ def upgrade() -> None:
         "md_refresh_token_status",
         sa.Column(
             "status_id",
-            sa.BigInteger(),
+            sa.SmallInteger(),
             sa.Identity(always=True),
             nullable=False,
         ),
